[INJIMOB-2810] move holder property to vp token instance creation area
