+++ Sources/OpenID4VP/AuthorizationResponse/VPTokenForSigning/VPTokenForSigning.swift
@@ -1,4 +1,5 @@
 import Foundation
 
+//TODO: rename to UnsignedVpToken
 public protocol VPTokenForSigning : Encodable {
 }
